Refactor screensaver X11 code to new file
