+++ internal/ui/desk.go
@@ -4,11 +4,9 @@ import (
 	"math"
 	"os/exec"
 	"strconv"
-	"time"
 
 	"fyshos.com/fynedesk/internal/notify"
 	"github.com/FyshOS/appie"
-	"github.com/FyshOS/saver"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/canvas"
@@ -364,56 +362,6 @@ func (l *desktop) registerShortcuts() {
 		func() {
 			l.TriggerScreensaver(true)
 		})
-}
-
-func (l *desktop) startXscreensaver() {
-	_, err := exec.LookPath("xscreensaver")
-	if err != nil {
-		fyne.LogError("xscreensaver command not found", err)
-		return
-	}
-	err = exec.Command("xscreensaver", "-no-splash").Start()
-	if err != nil {
-		fyne.LogError("Failed to lock screen", err)
-	}
-}
-
-func (l *desktop) TriggerScreensaver(lock bool) {
-	s := saver.NewScreenSaver(nil)
-	s.ClockFormat = l.settings.ClockFormatting()
-	if l.settings.ScreenSaverClock() {
-		s.Label = "(clock)"
-	} else {
-		s.Label = l.settings.ScreenSaverLabel()
-	}
-	s.Lock = lock
-
-	l.wm.ShowScreensaver(s)
-}
-
-var lastActivity time.Time
-
-func (l *desktop) DelayScreensaver() {
-	lastActivity = time.Now()
-}
-
-var idle bool
-
-func (l *desktop) watchScreenActivity() {
-	to := time.NewTicker(5 * time.Second)
-
-	for range to.C {
-		if lastActivity.Add(time.Minute * 5).Before(time.Now()) {
-
-			if !idle {
-				idle = true
-
-				l.TriggerScreensaver(false)
-			}
-		} else {
-			idle = false
-		}
-	}
 }
 
 // Screens returns the screens provider of the current desktop environment for access to screen functionality.
